Changed from uint to size_t

+++ src/something_game.hpp
@@ -112,7 +112,7 @@ struct Game
     void update(float dt);
     void render(SDL_Renderer *renderer);
     void handle_event(SDL_Event *event);
-    void render_debug_overlay(SDL_Renderer *renderer, Uint16 fps);
+    void render_debug_overlay(SDL_Renderer *renderer, size_t fps);
 
     // Entities of the Game
     void reset_entities();
