+++ src/something_main.cpp
@@ -155,8 +155,8 @@ int main(void)
     Uint32 prev_ticks = SDL_GetTicks();
     float lag_sec = 0;
     float next_sec = 0;
-    Uint32 frames_of_current_second = 0;
-    Uint16 fps = 0;
+    size_t frames_of_current_second = 0;
+    size_t fps = 0;
     while (!game.quit) {
         Uint32 curr_ticks = SDL_GetTicks();
         float elapsed_sec = (float) (curr_ticks - prev_ticks) / 1000.0f;
